+++ internal/block/block.go
@@ -4,8 +4,9 @@ import (
 	"bytes"
 	"encoding/gob"
 
-	"github.com/nspcc-dev/dbft/crypto"
-	"github.com/nspcc-dev/dbft/merkle"
+	"github.com/nspcc-dev/dbft"
+	"github.com/nspcc-dev/dbft/internal/crypto"
+	"github.com/nspcc-dev/dbft/internal/merkle"
 )
 
 type (
@@ -21,43 +22,11 @@ type (
 		NextConsensus crypto.Uint160
 	}
 
-	// Block is a generic interface for a block used by dbft.
-	Block[H crypto.Hash, A crypto.Address] interface {
-		// Hash returns block hash.
-		Hash() H
-
-		Version() uint32
-		// PrevHash returns previous block hash.
-		PrevHash() H
-		// MerkleRoot returns a merkle root of the transaction hashes.
-		MerkleRoot() H
-		// Timestamp returns block's proposal timestamp.
-		Timestamp() uint64
-		// Index returns block index.
-		Index() uint32
-		// ConsensusData is a random nonce.
-		ConsensusData() uint64
-		// NextConsensus returns hash of the validators of the next block.
-		NextConsensus() A
-
-		// Signature returns block's signature.
-		Signature() []byte
-		// Sign signs block and sets it's signature.
-		Sign(key crypto.PrivateKey) error
-		// Verify checks if signature is correct.
-		Verify(key crypto.PublicKey, sign []byte) error
-
-		// Transactions returns block's transaction list.
-		Transactions() []Transaction[H]
-		// SetTransactions sets block's transaction list.
-		SetTransactions([]Transaction[H])
-	}
-
 	neoBlock struct {
 		base
 
 		consensusData uint64
-		transactions  []Transaction[crypto.Uint256]
+		transactions  []dbft.Transaction[crypto.Uint256]
 		signature     []byte
 		hash          *crypto.Uint256
 	}
@@ -99,17 +68,17 @@ func (b *neoBlock) ConsensusData() uint64 {
 }
 
 // Transactions implements Block interface.
-func (b *neoBlock) Transactions() []Transaction[crypto.Uint256] {
+func (b *neoBlock) Transactions() []dbft.Transaction[crypto.Uint256] {
 	return b.transactions
 }
 
 // SetTransactions implements Block interface.
-func (b *neoBlock) SetTransactions(txx []Transaction[crypto.Uint256]) {
+func (b *neoBlock) SetTransactions(txx []dbft.Transaction[crypto.Uint256]) {
 	b.transactions = txx
 }
 
 // NewBlock returns new block.
-func NewBlock(timestamp uint64, index uint32, nextConsensus crypto.Uint160, prevHash crypto.Uint256, version uint32, nonce uint64, txHashes []crypto.Uint256) Block[crypto.Uint256, crypto.Uint160] {
+func NewBlock(timestamp uint64, index uint32, nextConsensus crypto.Uint160, prevHash crypto.Uint256, version uint32, nonce uint64, txHashes []crypto.Uint256) dbft.Block[crypto.Uint256, crypto.Uint160] {
 	block := new(neoBlock)
 	block.base.Timestamp = uint32(timestamp / 1000000000)
 	block.base.Index = index
@@ -144,7 +113,7 @@ func (b *neoBlock) GetHashData() []byte {
 }
 
 // Sign implements Block interface.
-func (b *neoBlock) Sign(key crypto.PrivateKey) error {
+func (b *neoBlock) Sign(key dbft.PrivateKey) error {
 	data := b.GetHashData()
 
 	sign, err := key.Sign(data)
@@ -158,7 +127,7 @@ func (b *neoBlock) Sign(key crypto.PrivateKey) error {
 }
 
 // Verify implements Block interface.
-func (b *neoBlock) Verify(pub crypto.PublicKey, sign []byte) error {
+func (b *neoBlock) Verify(pub dbft.PublicKey, sign []byte) error {
 	data := b.GetHashData()
 	return pub.Verify(data, sign)
 }