+++ internal/block/block_test.go
@@ -8,7 +8,8 @@ import (
 	"errors"
 	"testing"
 
-	"github.com/nspcc-dev/dbft/crypto"
+	"github.com/nspcc-dev/dbft"
+	"github.com/nspcc-dev/dbft/internal/crypto"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -18,7 +19,7 @@ func TestNeoBlock_Setters(t *testing.T) {
 
 	require.Equal(t, crypto.Uint256{}, b.Hash())
 
-	txs := []Transaction[crypto.Uint256]{testTx(1), testTx(2)}
+	txs := []dbft.Transaction[crypto.Uint256]{testTx(1), testTx(2)}
 	b.SetTransactions(txs)
 	assert.Equal(t, txs, b.Transactions())