+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Behaviour changes:
    (#94)
  * rename `InitializeConsensus` dBFT method to `Reset` (#95)
  * drop outdated dBFT `Service` interface (#95)
+ * move all default implementations to `internal` package (#97)
 
 Improvements:
  * add MIT License (#78, #79)